Moved mocks generator functions to mock.ts file
